Routes should work. Added ThemeProvider to app.js. I moved components with styles to separate files.
