+++ src/Context/styles.js
@@ -1,0 +1,15 @@
+export const theme= {
+   colors: {
+     hover: '#d1c4e9',
+     clickedNav: '#aa90d7',
+     grey: '#d3d4d5',
+     black: '#3e3e3e',
+   },
+   fontSize: {
+     xl: '24px',
+     l: '16px',
+     m: '12px',
+     s: '11px',
+   },
+ };
+
