+++ src/components/Atoms/Label/Label.js
@@ -3,6 +3,6 @@ import styled from 'styled-components';
 export const Label = styled.label`
   font-family: Montserrat, sans-serif;
   font-weight: bold;
-  font-size: 16px;
-  color: #3e3e3e;
+  font-size: ${({theme}) => theme.fontSize.l};
+  color: ${({theme}) => theme.colors.black};
 `;
