+++ src/components/Atoms/NavigationText/NavigationText.js
@@ -3,15 +3,14 @@ import Button from '@material-ui/core/Button';
 import { withStyles } from '@material-ui/core/styles';
 
 
-const StyledButton = withStyles((theme) => ({
+const StyledButton = withStyles(() => ({
   root: {
     '&:hover': {
-      transition: '.6s ease-in',
+      transition: '.3s ease-in',
       backgroundColor: '#d1c4e9',
     },
   },
 }))(Button);
-
 const NavigationText = ({ mainText, variant = 'text', handleClick = null }) => {
   return (
     <>
@@ -20,7 +19,7 @@ const NavigationText = ({ mainText, variant = 'text', handleClick = null }) => {
         aria-haspopup="true"
         size="large"
         variant={variant}
-        onMouseOver={handleClick}
+        onClick={handleClick}
       >
         {mainText}
       </StyledButton>
