+++ src/components/Loaders/CircularLoader/CircularLoader.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import CircularProgress from '@material-ui/core/CircularProgress';
+
+
+
+export default function CircularIndeterminate() {
+  return (
+    <div>
+      <CircularProgress color="secondary" size={30} />
+    </div>
+  );
+}
