+++ src/components/Molecules/PaymentFormField/PaymentFormField.js
@@ -1,18 +1,8 @@
 import { TextField } from '@material-ui/core';
 import React from 'react';
-import styled from 'styled-components';
 import { Label } from '../../Atoms/Label/Label';
+import {Wrapper} from './PaymentFormFieldWrapper.styles'
 
-export const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 30px;
-  ${Label} {
-    margin: 10px auto;
-  }
-`;
 
 const PaymentFormField = ({
   label,
