+++ src/view/App.js
@@ -1,10 +1,15 @@
 import React from 'react';
+import { ThemeProvider } from 'styled-components';
 import NavigationView from '../components/templates/NavigationTemplate/NavigationTemplate';
-
+import { theme } from '../Context/styles';
 
 
 function App() {
-  return <NavigationView />;
+  return (
+    <ThemeProvider theme={theme}>
+      <NavigationView />
+    </ThemeProvider>
+  );
 }
 
 export default App;
